requirements.txt
mecab-python3
slackbot
